third_party/WebKit/LayoutTests/virtual/service-worker-script-streaming/external/wpt/service-workers/service-worker/ServiceWorkerGlobalScope/registration-attribute.https-expected.txt

This is a testharness.js-based test.
PASS Verify registration attributes on ServiceWorkerGlobalScope
FAIL Verify registration attributes on ServiceWorkerGlobalScope of the newer worker assert_array_equals: property 12, expected "statechange(installed)" but got "activate"
Harness: the test ran to completion.
